correctly juggle buffer flushing on win32
The problem:  if you write() to a file and then read() from it without
an intervening flush(), win32 will happily return junk data from beyond
the end of the file.

The solution:  only request as much data as we expect to be in the
buffer, and flush the buffer after consuming the whole input stream.